requirements.txt
torch
torchvision

opencv-python
Pillow
numpy